chore: bump notify on second time
