+++ .dumi/theme/slots/Header/index.tsx
@@ -44,7 +44,7 @@ const useStyle = () => {
       }
 
       .dumi-default-search-bar {
-        border-inline-start: 1px solid rgba(0,0,0,.06);
+        border-inline-start: 1px solid rgba(0, 0, 0, 0.06);
 
         > svg {
           width: 14px;
@@ -135,16 +135,6 @@ const triggerDocSearchImport = () => {
   });
 };
 
-const SHOULD_OPEN_ANT_DESIGN_MIRROR_MODAL = 'ANT_DESIGN_DO_NOT_OPEN_MIRROR_MODAL';
-
-function disableAntdMirrorModal() {
-  window.localStorage.setItem(SHOULD_OPEN_ANT_DESIGN_MIRROR_MODAL, 'true');
-}
-
-function shouldOpenAntdMirrorModal() {
-  return !window.localStorage.getItem(SHOULD_OPEN_ANT_DESIGN_MIRROR_MODAL);
-}
-
 interface HeaderState {
   menuVisible: boolean;
   windowWidth: number;
@@ -203,26 +193,6 @@ const Header: React.FC<HeaderProps> = (props) => {
     pingTimer.current = ping((status) => {
       if (status !== 'timeout' && status !== 'error') {
         setHeaderState((prev) => ({ ...prev, showTechUIButton: true }));
-        if (
-          process.env.NODE_ENV === 'production' &&
-          window.location.host !== 'ant-design.antgroup.com' &&
-          shouldOpenAntdMirrorModal()
-        ) {
-          Modal.confirm({
-            title: '提示',
-            content: '内网用户推荐访问国内镜像以获得极速体验～',
-            okText: '🚀 立刻前往',
-            cancelText: '不再弹出',
-            closable: true,
-            onOk() {
-              window.open('https://ant-design.antgroup.com', '_self');
-              disableAntdMirrorModal();
-            },
-            onCancel() {
-              disableAntdMirrorModal();
-            },
-          });
-        }
       }
     });
     return () => {
